+++ BlueprintUILists/Sources/Element+HeaderFooter.swift
@@ -6,46 +6,10 @@
 //
 
 import BlueprintUI
-@_spi(ListableInternal)
 import ListableUI
 
 
 // MARK: HeaderFooter / HeaderFooterContent Extensions
-
-
-extension Element {
-        
-    /// Converts the given `Element` into a Listable `HeaderFooter`. You many also optionally
-    /// configure the header / footer, setting its values such as the `onTap` callbacks, etc.
-    ///
-    /// You may also provide a background and pressed background as well as tap actions.
-    ///
-    /// ```swift
-    /// MyElement(...)
-    ///     .listHeaderFooter { header in
-    ///         header.onTap = { ... }
-    ///     } background: {
-    ///         Box(backgroundColor: ...).inset(...)
-    ///     } pressedBackground: {
-    ///         Box(backgroundColor: ...).inset(...)
-    ///     }
-    /// ```
-    public func listHeaderFooter(
-        background : @escaping () -> Element? = { nil },
-        pressedBackground : @escaping () -> Element? = { nil },
-        configure : (inout HeaderFooter<WrappedHeaderFooterContent<Self>>) -> () = { _ in }
-    ) -> HeaderFooter<WrappedHeaderFooterContent<Self>> {
-        HeaderFooter(
-            WrappedHeaderFooterContent(
-                represented: self,
-                background: background,
-                pressedBackground: pressedBackground
-            ),
-            configure: configure
-        )
-    }
-}
-
 
 
 /// Ensures that the `Equatable` initializer for `WrappedHeaderFooterContent` is called.
@@ -68,9 +32,10 @@ extension Element where Self:Equatable {
 }
 
 
-/// Ensures that the `EquivalentComparable` initializer for `WrappedHeaderFooterContent` is called.
-extension Element where Self:EquivalentComparable {
+/// Ensures that the `LayoutEquivalent` initializer for `WrappedHeaderFooterContent` is called.
+extension Element where Self:LayoutEquivalent {
     
+    @_disfavoredOverload
     public func listHeaderFooter(
         background : @escaping () -> Element? = { nil },
         pressedBackground : @escaping () -> Element? = { nil },
@@ -98,21 +63,6 @@ public struct WrappedHeaderFooterContent<ElementType:Element> : BlueprintHeaderF
         represented : ElementType,
         background : @escaping () -> Element?,
         pressedBackground : @escaping () -> Element?
-    ) {
-        self.represented = represented
-        
-        self.backgroundProvider = background
-        self.pressedBackgroundProvider = pressedBackground
-        
-        self.isEquivalent = {
-            defaultIsEquivalentImplementation($0.represented, $1.represented)
-        }
-    }
-    
-    init(
-        represented : ElementType,
-        background : @escaping () -> Element?,
-        pressedBackground : @escaping () -> Element?
     ) where ElementType:Equatable
     {
         self.represented = represented
@@ -129,7 +79,7 @@ public struct WrappedHeaderFooterContent<ElementType:Element> : BlueprintHeaderF
         represented : ElementType,
         background : @escaping () -> Element?,
         pressedBackground : @escaping () -> Element?
-    ) where ElementType:EquivalentComparable
+    ) where ElementType:LayoutEquivalent
     {
         self.represented = represented