+++ BlueprintUILists/Sources/Section+Element.swift
@@ -19,19 +19,35 @@ extension Section {
     ///     section.add(Element2())
     /// }
     /// ```
-    public mutating func add<ElementType:Element>(_ element : ElementType)
-    {
-        self.items.append(element.listItem())
-    }
-    
     public mutating func add<ElementType:Element>(_ element : ElementType) where ElementType:Equatable
     {
         self.items.append(element.listItem())
     }
     
-    public mutating func add<ElementType:Element>(_ element : ElementType) where ElementType:EquivalentComparable
+    /// Adds `Element` support when building a `Section`.
+    ///
+    /// ```swift
+    /// Section("id") { section in
+    ///     section.add(Element1())
+    ///     section.add(Element2())
+    /// }
+    /// ```
+    public mutating func add<ElementType:Element>(_ element : ElementType) where ElementType:LayoutEquivalent
     {
         self.items.append(element.listItem())
+    }
+        
+    /// Adds `Element` support when building a `Section`.
+    ///
+    /// ```swift
+    /// Section("id") { section in
+    ///     section += Element1()
+    ///     section += Element2()
+    /// }
+    /// ```
+    public static func += <ElementType:Element>(lhs : inout Section, rhs : ElementType) where ElementType:Equatable
+    {
+        lhs.add(rhs)
     }
     
     /// Adds `Element` support when building a `Section`.
@@ -42,17 +58,8 @@ extension Section {
     ///     section += Element2()
     /// }
     /// ```
-    public static func += <ElementType:Element>(lhs : inout Section, rhs : ElementType)
-    {
-        lhs.add(rhs)
-    }
-    
-    public static func += <ElementType:Element>(lhs : inout Section, rhs : ElementType) where ElementType:Equatable
-    {
-        lhs.add(rhs)
-    }
-    
-    public static func += <ElementType:Element>(lhs : inout Section, rhs : ElementType) where ElementType:EquivalentComparable
+    @_disfavoredOverload
+    public static func += <ElementType:Element>(lhs : inout Section, rhs : ElementType) where ElementType:LayoutEquivalent
     {
         lhs.add(rhs)
     }